apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
supabase==2.0.2
python-dotenv==1.0.0
pandas==2.1.3
bcrypt==4.1.1